LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_null-expected.txt

Description

This test validates that for a cross origin resource, the timing allow check algorithm will fail when the value of Timing-Allow-Origin is null.


Harness Error (FAIL), message = Error: assert_greater_than: The iframe should have at least one resource timing entry. expected a number greater than 0 but got 0

PASS window.performance is defined
PASS window.performance.getEntriesByType is defined
